feat(Tool): Implementing majority of Tool Class

+++ src/Consumer/Advantage/Classes/Tool.js
@@ -14,31 +14,30 @@ const validScopes = require('../../../GlobalUtils/Helpers/scopes')
  * @description Class representing a registered tool.
  */
 class Tool {
-  
   #clientId
 
   #deploymentId
-  
+
   #url
-  
+
   #deepLinkingUrl
 
   #loginUrl
-  
+
   #redirectionURIs
-  
+
   #name
-  
+
   #description
-  
+
   #authConfig
 
   #scopes = []
 
   #privacy
-  
+
   #customParameters = {}
-  
+
   #kid
 
   /**
@@ -79,14 +78,13 @@ class Tool {
    * @returns {Promise<Tool | false>}
    */
   static async getTool (clientId) {
-    if (!url) throw new Error('MISSING_CLIENT_ID_PARAMETER')
+    if (!clientId) throw new Error('MISSING_CLIENT_ID_PARAMETER')
     const result = await Database.get('tool', { clientId: clientId })
     if (!result) return false
     const _tool = result[0]
     const tool = new Tool(clientId, _tool.deploymentId, _tool.url, _tool.deepLinkingUrl, _tool.loginUrl, _tool.redirectionURIs, _tool.name, _tool.description, _tool.authConfig, _tool.scopes, _tool.privacy, _tool.customParameters, _tool.kid)
     return tool
   }
-
 
   /**
    * @description Gets all tools.
@@ -96,7 +94,7 @@ class Tool {
     const result = []
     const tools = await Database.get('tool')
     if (tools) {
-      for (const _tool of tools) result.push(new Tool(clientId, _tool.deploymentId, _tool.url, _tool.deepLinkingUrl, _tool.loginUrl, _tool.redirectionURIs, _tool.name, _tool.description, _tool.authConfig, _tool.scopes, _tool.privacy, _tool.customParameters, _tool.kid))
+      for (const _tool of tools) result.push(new Tool(_tool.clientId, _tool.deploymentId, _tool.url, _tool.deepLinkingUrl, _tool.loginUrl, _tool.redirectionURIs, _tool.name, _tool.description, _tool.authConfig, _tool.scopes, _tool.privacy, _tool.customParameters, _tool.kid))
     }
     return result
   }
@@ -106,10 +104,10 @@ class Tool {
    * @param {Object} tool - Tool configuration object.
    * @param {string} tool.url - Tool url.
    * @param {string} tool.name - Tool name.
-   * @param {string} tool.deepLinkingUrl - Tool deep linking url.
    * @param {string} tool.loginUrl - Tool login url.
-   * @param {string} tool.redirectionURIs - Tool redirection URIs.
    * @param {Object} tool.authConfig - Authentication configurations for the tool.
+   * @param {string} [tool.redirectionURIs] - Tool redirection URIs.
+   * @param {string} [tool.deepLinkingUrl] - Tool deep linking url.
    * @param {string} [tool.clientId] - Tool Client Id.
    * @param {string} [tool.description] - Tool description.
    * @param {Array<String>} [tool.scopes] - Scopes allowed for the tool.
@@ -118,45 +116,50 @@ class Tool {
    * @returns {Promise<Tool>}
    */
   static async registerTool (tool) {
-    if (!tool || !tool.url || !tool.name || !tool.deepLinkingUrl || !tool.loginUrl || !tool.redirectionURIs || !tool.authConfig) throw new Error('MISSING_REGISTRATION_PARAMETERS')
-
+    if (!tool || !tool.url || !tool.name || !tool.loginUrl || !tool.authConfig) throw new Error('MISSING_REGISTRATION_PARAMETERS')
     if (tool.authConfig.method !== 'RSA_KEY' && tool.authConfig.method !== 'JWK_KEY' && tool.authConfig.method !== 'JWK_SET') throw new Error('INVALID_AUTHCONFIG_METHOD. Details: Valid methods are "RSA_KEY", "JWK_KEY", "JWK_SET".')
     if (!tool.authConfig.key) throw new Error('MISSING_AUTHCONFIG_KEY')
 
-    let clientId = tool.clientId
-    if (clientId) {
-      const tool = await Tool.getTool(tool.clientId)
-      if (tool) throw new Error('TOOL_ALREADY_REGISTERED')
-    } else {
-      clientId = uuidv4()
-      while (await Database.get('tool', { clientId: clientId })) {
-        clientId = uuidv4()
-      }
-    }
-    const deploymentId = uuidv4()
-    while (await Database.get('tool', { deploymentId: deploymentId })) {
-      deploymentId = uuidv4()
-    }
+    if (!tool.description) tool.description = ''
+    if (!tool.deepLinkingUrl) tool.deepLinkingUrl = tool.url
 
-    const privacy = {
-      name: (tool.privacy && tool.privacy.name === true) ? true : false,
-      email: (tool.privacy && tool.privacy.email === true) ? true : false
-    }
-    const customParameters = tool.customParameters || {}
+    if (!tool.redirectionURIs) tool.redirectionURIs = []
+    else if (!Array.isArray(tool.redirectionURIs)) throw new Error('INVALID_REDIRECTION_URIS_ARRAY')
 
-    let scopes
-    if (tool.scopes) {
+    if (!tool.customParameters) tool.customParameters = {}
+    else if (typeof tool.customParameters !== 'object') throw new Error('INVALID_CUSTOM_PARAMETERS_OBJECT')
+
+    if (!tool.scopes) tool.scopes = []
+    else {
+      if (!Array.isArray(tool.scopes)) throw new Error('INVALID_SCOPES_ARRAY')
       for (const scope of tool.scopes) {
         if (!Object.keys(validScopes).includes(scope)) throw new Error('INVALID_SCOPE. Details: Invalid scope: ' + scope)
       }
-      scopes = tool.scopes
-    } else scopes = []
+    }
+
+    tool.privacy = {
+      name: (tool.privacy && tool.privacy.name === true),
+      email: (tool.privacy && tool.privacy.email === true)
+    }
+
+    if (tool.clientId) {
+      const _tool = await Tool.getTool(tool.clientId)
+      if (_tool) throw new Error('TOOL_ALREADY_REGISTERED')
+    } else {
+      tool.clientId = uuidv4()
+      while (await Database.get('tool', { clientId: tool.clientId })) {
+        tool.clientId = uuidv4()
+      }
+    }
+    tool.deploymentId = uuidv4()
+    while (await Database.get('tool', { deploymentId: tool.deploymentId })) {
+      tool.deploymentId = uuidv4()
+    }
 
     let kid
     try {
       consToolDebug('Registering new tool')
-      consToolDebug('Tool ClientId: ' + clientId)
-
+      consToolDebug('Tool Client ID: ' + tool.clientId)
       // Generating and storing RSA keys
       const keyPair = await Keyset.generateKeyPair()
       kid = keyPair.kid
@@ -164,13 +167,15 @@ class Tool {
       await Database.replace('privatekey', { clientId: tool.clientId }, { key: keyPair.privateKey, kid: kid }, true, { kid: kid, clientId: tool.clientId })
 
       // Storing new tool
-      await Database.replace('tool', { toolUrl: tool.url, clientId: tool.clientId }, { toolName: tool.name, toolUrl: tool.url, clientId: tool.clientId, authEndpoint: tool.authenticationEndpoint, accesstokenEndpoint: tool.accesstokenEndpoint, kid: kid, authConfig: tool.authConfig })
+      await Database.replace('tool', { clientId: tool.clientId }, { clientId: tool.clientId, deploymentId: tool.deploymentId, url: tool.url, deepLinkingUrl: tool.deepLinkingUrl, loginUrl: tool.loginUrl, redirectionURIs: tool.redirectionURIs, name: tool.name, description: tool.description, authConfig: tool.authConfig, scopes: tool.scopes, privacy: tool.privacy, customParameters: tool.customParameters, kid: kid })
 
-      const plat = new Tool(kid, tool.name, tool.url, tool.clientId, tool.authenticationEndpoint, tool.accesstokenEndpoint, tool.authConfig)
-      return plat
+      const _tool = new Tool(tool.clientId, tool.deploymentId, tool.url, tool.deepLinkingUrl, tool.loginUrl, tool.redirectionURIs, tool.name, tool.description, tool.authConfig, tool.scopes, tool.privacy, tool.customParameters, kid)
+      return _tool
     } catch (err) {
-      await Database.delete('publickey', { kid: kid })
-      await Database.delete('privatekey', { kid: kid })
+      if (kid) {
+        await Database.delete('publickey', { kid: kid })
+        await Database.delete('privatekey', { kid: kid })
+      }
       await Database.delete('tool', { clientId: tool.clientId })
       consToolDebug(err.message)
       throw (err)
@@ -179,67 +184,46 @@ class Tool {
 
   /**
    * @description Updates a tool by the Id.
-   * @param {String} toolId - Tool Id.
-   * @param {Object} toolInfo - Update Information.
-   * @param {String} toolInfo.url - Tool url.
-   * @param {String} toolInfo.clientId - Tool clientId.
-   * @param {String} toolInfo.name - Tool nickname.
-   * @param {String} toolInfo.authenticationEndpoint - Authentication endpoint that the tool will use to authenticate within the tool.
-   * @param {String} toolInfo.accesstokenEndpoint - Access token endpoint that the tool will use to get an access token for the tool.
-   * @param {object} toolInfo.authConfig - Authentication method and key for verifying messages from the tool. {method: "RSA_KEY", key:"PUBLIC KEY..."}
-   * @param {String} toolInfo.authConfig.method - Method of authorization "RSA_KEY" or "JWK_KEY" or "JWK_SET".
-   * @param {String} toolInfo.authConfig.key - Either the RSA public key provided by the tool, or the JWK key, or the JWK keyset address.
+   * @param {String} clientId - Tool Client ID.
+   * @param {string} toolInfo.url - Tool url.
+   * @param {string} toolInfo.name - Tool name.
+   * @param {string} toolInfo.loginUrl - Tool login url.
+   * @param {Object} toolInfo.authConfig - Authentication configurations for the tool.
+   * @param {string} toolInfo.redirectionURIs - Tool redirection URIs.
+   * @param {string} toolInfo.deepLinkingUrl - Tool deep linking url.
+   * @param {string} toolInfo.description - Tool description.
+   * @param {Array<String>} toolInfo.scopes - Scopes allowed for the tool.
+   * @param {Object} toolInfo.privacy - Privacy configuration.
+   * @param {Object} tool.customParameters - Globally set custom parameters.
    * @returns {Promise<Tool | false>}
    */
-  static async updateToolById (toolId, toolInfo) {
-    if (!toolId) { throw new Error('MISSING_PLATFORM_ID') }
+  static async updateTool (clientId, toolInfo) {
+    if (!clientId) { throw new Error('MISSING_PLATFORM_ID') }
     if (!toolInfo) { throw new Error('MISSING_PLATFORM_INFO') }
 
-    const tool = await Tool.getToolById(toolId)
-    if (!tool) return false
-
-    const oldURL = await tool.toolUrl()
-    const oldClientId = await tool.toolClientId()
+    const toolObject = await Tool.getTool(clientId)
+    if (!toolObject) return false
+    const tool = await toolObject.toJSON()
 
     const update = {
-      url: toolInfo.url || oldURL,
-      clientId: toolInfo.clientId || oldClientId,
-      name: toolInfo.name || await tool.toolName(),
-      authenticationEndpoint: toolInfo.authenticationEndpoint || await tool.toolAuthenticationEndpoint(),
-      accesstokenEndpoint: toolInfo.accesstokenEndpoint || await tool.toolAccessTokenEndpoint()
-    }
-
-    const authConfig = await tool.toolAuthConfig()
-    update.authConfig = authConfig
-    if (toolInfo.authConfig) {
-      if (toolInfo.authConfig.method) update.authConfig.method = toolInfo.authConfig.method
-      if (toolInfo.authConfig.key) update.authConfig.key = toolInfo.authConfig.key
-    }
-
-    let alteredUrlClientIdFlag = false
-    if (toolInfo.url || toolInfo.clientId) {
-      if (toolInfo.url !== oldURL || toolInfo.clientId !== oldClientId) alteredUrlClientIdFlag = true
-    }
-
-    if (alteredUrlClientIdFlag) {
-      if (await Database.get('tool', { toolUrl: update.url, clientId: update.clientId })) throw new Error('URL_CLIENT_ID_COMBINATION_ALREADY_EXISTS')
+      url: toolInfo.url || tool.url,
+      deepLinkingUrl: toolInfo.deepLinkingUrl || tool.deepLinkingUrl,
+      loginUrl: toolInfo.loginUr || tool.loginUrl,
+      redirectionURIs: toolInfo.redirectionURIs || tool.redirectionURIs,
+      name: toolInfo.name || tool.name,
+      description: toolInfo.description || tool.description,
+      authConfig: toolInfo.authConfig || tool.authConfig,
+      scopes: toolInfo.scopes || tool.scopes,
+      privacy: toolInfo.privacy || tool.privacy,
+      customParameters: toolInfo.customParameters || tool.customParameters
     }
 
     try {
-      if (alteredUrlClientIdFlag) {
-        await Database.modify('publickey', { kid: toolId }, { toolUrl: update.url, clientId: update.clientId })
-        await Database.modify('privatekey', { kid: toolId }, { toolUrl: update.url, clientId: update.clientId })
-      }
+      await Database.modify('tool', { clientId: clientId }, update)
 
-      await Database.modify('tool', { kid: toolId }, { toolUrl: update.url, clientId: update.clientId, toolName: update.name, authEndpoint: update.authenticationEndpoint, accesstokenEndpoint: update.accesstokenEndpoint, authConfig: update.authConfig })
-
-      const tool = new Tool(toolId, update.name, update.url, update.clientId, update.authenticationEndpoint, update.accesstokenEndpoint, update.authConfig)
-      return tool
+      const _tool = new Tool(clientId, tool.deploymentId, update.url, update.deepLinkingUrl, update.loginUrl, update.redirectionURIs, update.name, update.description, update.authConfig, update.scopes, update.privacy, update.customParameters, tool.kid)
+      return _tool
     } catch (err) {
-      if (alteredUrlClientIdFlag) {
-        await Database.modify('publickey', { kid: toolId }, { toolUrl: oldURL, clientId: oldClientId })
-        await Database.modify('privatekey', { kid: toolId }, { toolUrl: oldURL, clientId: oldClientId })
-      }
       consToolDebug(err.message)
       throw (err)
     }
@@ -259,129 +243,65 @@ class Tool {
 
   // Instance methods
   /**
-   * @description Gets the tool url.
-   */
-  async toolUrl () {
-    return this.#toolUrl
-  }
-
-  /**
    * @description Gets the tool client id.
    */
-  async toolClientId () {
+  async clientId () {
     return this.#clientId
   }
 
   /**
-     * @description Sets/Gets the tool name.
-     * @param {string} [name] - Tool name.
-     */
-  async toolName (name) {
-    if (!name) return this.#toolName
-    await Database.modify('tool', { toolUrl: this.#toolUrl, clientId: this.#clientId }, { toolName: name })
-    this.#toolName = name
+   * @description Gets/Sets the tool url.
+   */
+  async url () {
+    return this.#url
+  }
+
+  /**
+   * @description Sets/Gets the tool name.
+   * @param {string} [name] - Tool name.
+   */
+  async name (name) {
+    if (!name) return this.#name
+    await Database.modify('tool', { clientId: this.#clientId }, { name: name })
+    this.#name = name
     return name
   }
 
   /**
-     * @description Gets the tool Id.
-     */
-  async toolId () {
+   * @description Gets the tool key id.
+   */
+  async kid () {
     return this.#kid
   }
 
   /**
-   * @description Gets the tool key_id.
+   * @description Gets the RSA public key assigned to the tool.
+   *
    */
-  async toolKid () {
-    return this.#kid
-  }
-
-  /**
-   * @description Sets/Gets the tool status.
-   * @param {Boolean} [active] - Whether the Tool is active or not.
-   */
-  async toolActive (active) {
-    if (active === undefined) {
-      // Get tool status
-      const toolStatus = await Database.get('toolStatus', { id: this.#kid })
-      if (!toolStatus || toolStatus[0].active) return true
-      else return false
-    }
-    await Database.replace('toolStatus', { id: this.#kid }, { id: this.#kid, active: active })
-    return active
-  }
-
-  /**
-     * @description Gets the RSA public key assigned to the tool.
-     *
-     */
-  async toolPublicKey () {
+  async publicKey () {
     const key = await Database.get('publickey', { kid: this.#kid }, true)
     return key[0].key
   }
 
   /**
-     * @description Gets the RSA private key assigned to the tool.
-     *
-     */
-  async toolPrivateKey () {
+   * @description Gets the RSA private key assigned to the tool.
+   *
+   */
+  async privateKey () {
     const key = await Database.get('privatekey', { kid: this.#kid }, true)
     return key[0].key
   }
 
   /**
-     * @description Sets/Gets the tool authorization configurations used to validate it's messages.
-     * @param {string} method - Method of authorization "RSA_KEY" or "JWK_KEY" or "JWK_SET".
-     * @param {string} key - Either the RSA public key provided by the tool, or the JWK key, or the JWK keyset address.
-     */
-  async toolAuthConfig (method, key) {
-    if (!method && !key) return this.#authConfig
-
-    if (method && method !== 'RSA_KEY' && method !== 'JWK_KEY' && method !== 'JWK_SET') throw new Error('INVALID_METHOD. Details: Valid methods are "RSA_KEY", "JWK_KEY", "JWK_SET".')
-
-    const authConfig = {
-      method: method || this.#authConfig.method,
-      key: key || this.#authConfig.key
-    }
-
-    await Database.modify('tool', { toolUrl: this.#toolUrl, clientId: this.#clientId }, { authConfig: authConfig })
-    this.#authConfig = authConfig
-    return authConfig
-  }
-
-  /**
-   * @description Sets/Gets the tool authorization endpoint used to perform the OIDC login.
-   * @param {string} [authenticationEndpoint - Tool authentication endpoint.
+   * @description Gets the tool access token or attempts to generate a new one.
+   * @param {String} scopes - String of scopes.
    */
-  async toolAuthenticationEndpoint (authenticationEndpoint) {
-    if (!authenticationEndpoint) return this.#authenticationEndpoint
-    await Database.modify('tool', { toolUrl: this.#toolUrl, clientId: this.#clientId }, { authEndpoint: authenticationEndpoint })
-    this.#authenticationEndpoint = authenticationEndpoint
-    return authenticationEndpoint
-  }
-
-  /**
-     * @description Sets/Gets the tool access token endpoint used to authenticate messages to the tool.
-     * @param {string} [accesstokenEndpoint] - Tool access token endpoint.
-     */
-  async toolAccessTokenEndpoint (accesstokenEndpoint) {
-    if (!accesstokenEndpoint) return this.#accesstokenEndpoint
-    await Database.modify('tool', { toolUrl: this.#toolUrl, clientId: this.#clientId }, { accesstokenEndpoint: accesstokenEndpoint })
-    this.#accesstokenEndpoint = accesstokenEndpoint
-    return accesstokenEndpoint
-  }
-
-  /**
-     * @description Gets the tool access token or attempts to generate a new one.
-     * @param {String} scopes - String of scopes.
-     */
-  async toolAccessToken (scopes) {
-    const result = await Database.get('accesstoken', { toolUrl: this.#toolUrl, clientId: this.#clientId, scopes: scopes }, true)
+  async accessToken (scopes) {
+    const result = await Database.get('accesstoken', { toolUrl: this.#url, clientId: this.#clientId, scopes: scopes }, true)
     let token
     if (!result || (Date.now() - result[0].createdAt) / 1000 > result[0].token.expires_in) {
-      consToolDebug('Valid access_token for ' + this.#toolUrl + ' not found')
-      consToolDebug('Attempting to generate new access_token for ' + this.#toolUrl)
+      consToolDebug('Valid access_token for ' + this.#url + ' not found')
+      consToolDebug('Attempting to generate new access_token for ' + this.#url)
       consToolDebug('With scopes: ' + scopes)
       token = await Auth.generateAccessToken(scopes, this)
     } else {
@@ -395,9 +315,9 @@ class Tool {
   /**
    * @description Retrieves the tool information as a JSON object.
    */
-  async toolJSON () {
+  async toJSON () {
     const toolJSON = {
-      id: this.#kid,
+      kid: this.#kid,
       url: this.#url,
       clientId: this.#clientId,
       deploymentId: this.#deploymentId,
